# docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Interview API
  version: 1.0.0
  description: Frontend contract verified against the current Spring Boot implementation.
servers:
  - url: http://localhost:8080
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '201':
          description: Registered; also sets the refresh-token cookie.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Logged in; also sets the refresh-token cookie.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/auth/google:
    post:
      tags: [Auth]
      operationId: loginWithGoogle
      summary: Log in with a Google ID token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GoogleLoginRequest'}
      responses:
        '200':
          description: Logged in; also sets the refresh-token cookie.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the current user
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentUserResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshAccessToken
      summary: Rotate the refresh token
      security:
        - refreshCookie: []
      responses:
        '200':
          description: Rotated refresh token and new access token.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/auth/logout-all:
    post:
      tags: [Auth]
      operationId: logoutAll
      summary: Log out all devices
      responses:
        '204': {description: All refresh tokens revoked.}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/profiles:
    get:
      tags: [CV and Profile]
      operationId: listProfiles
      summary: List candidate profiles
      responses:
        '200':
          description: Profiles backed by active CVs.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ProfileSummary'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/profiles/{profileId}:
    parameters:
      - {$ref: '#/components/parameters/ProfileId'}
    get:
      tags: [CV and Profile]
      operationId: getProfile
      summary: Get a candidate profile
      responses:
        '200':
          description: Candidate profile.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CandidateProfile'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [CV and Profile]
      operationId: updateProfile
      summary: Replace a candidate profile
      description: Full replacement of scalars and all three child collections.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProfileUpdateRequest'}
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CandidateProfile'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/profiles/{profileId}/confirm:
    parameters:
      - {$ref: '#/components/parameters/ProfileId'}
    post:
      tags: [CV and Profile]
      operationId: confirmProfile
      summary: Confirm a candidate profile
      responses:
        '200':
          description: Confirmed profile; the operation is idempotent.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CandidateProfile'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
...
    ProfileId:
      name: profileId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    BadRequest:
      description: Invalid input.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Unauthorized:
      description: Missing or invalid authentication.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Forbidden:
      description: Authenticated but not allowed.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    NotFound:
      description: Resource not found or inaccessible.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Conflict:
      description: State, idempotency, version, or limit conflict.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
    ServiceUnavailable:
      description: AI, storage, or configuration unavailable.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, code, message, path]
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer}
        code: {type: string}
        message: {type: string}
        path: {type: string}
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: {type: string}
    UserRole: {type: string, enum: [ADMIN, USER]}
    RegisterRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName: {type: string, minLength: 1, maxLength: 150}
        email: {type: string, format: email, minLength: 1, maxLength: 150}
        password: {type: string, minLength: 8, maxLength: 100, writeOnly: true}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 1, writeOnly: true}
    GoogleLoginRequest:
      type: object
      required: [idToken]
      properties:
        idToken: {type: string, minLength: 1, writeOnly: true}
    AuthResponse:
      type: object
      required: [accessToken, userId, email, role]
      properties:
        accessToken: {type: string}
        userId: {type: integer, format: int64}
        email: {type: string, format: email}
        role: {$ref: '#/components/schemas/UserRole'}
    CurrentUserResponse:
      type: object
      required: [id, fullName, email, role, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64}
        fullName: {type: string}
        email: {type: string, format: email}
        avatarUrl: {type: string, format: uri, nullable: true}
        role: {$ref: '#/components/schemas/UserRole'}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
...
    ProfileEducation:
      type: object
      required: [school]
      properties:
        id: {type: integer, format: int64, minimum: 1, nullable: true}
        school: {type: string, minLength: 1, maxLength: 255}
        degree: {type: string, maxLength: 150, nullable: true}
        fieldOfStudy: {type: string, maxLength: 150, nullable: true}
        startYear: {type: integer, minimum: 1900, maximum: 2100, nullable: true}
        endYear: {type: integer, minimum: 1900, maximum: 2100, nullable: true}
        userEdited: {type: boolean, nullable: true, readOnly: true}
        displayOrder: {type: integer, nullable: true, readOnly: true}
    ProfileSkill:
      type: object
      required: [name]
      properties:
        id: {type: integer, format: int64, minimum: 1, nullable: true}
        name: {type: string, minLength: 1, maxLength: 80}
        category: {type: string, maxLength: 50, nullable: true}
        userEdited: {type: boolean, nullable: true, readOnly: true}
        displayOrder: {type: integer, nullable: true, readOnly: true}
    ProfileProject:
      type: object
      required: [name]
      properties:
        id: {type: integer, format: int64, minimum: 1, nullable: true}
        name: {type: string, minLength: 1, maxLength: 255}
        description: {type: string, maxLength: 5000, nullable: true}
        roleInProject: {type: string, maxLength: 150, nullable: true}
        techStack: {type: string, maxLength: 500, nullable: true}
        startDate: {type: string, format: date, nullable: true}
        endDate: {type: string, format: date, nullable: true}
        userEdited: {type: boolean, nullable: true, readOnly: true}
        displayOrder: {type: integer, nullable: true, readOnly: true}
    ProfileSummary:
      type: object
      required: [id, version, name, cvDocumentId, cvOriginalFilename, source, educationCount, skillCount, projectCount, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64}
        version: {type: integer, format: int64}
        name: {type: string}
        cvDocumentId: {type: integer, format: int64}
        cvOriginalFilename: {type: string}
        headline: {type: string, nullable: true}
        targetPosition: {type: string, nullable: true}
        seniorityLevel: {type: string, nullable: true}
        source: {type: string, enum: [AUTO_PARSED, USER_EDITED]}
        confirmedAt: {type: string, format: date-time, nullable: true}
        educationCount: {type: integer}
        skillCount: {type: integer}
        projectCount: {type: integer}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    CandidateProfile:
      type: object
      required: [id, version, name, cvDocumentId, cvOriginalFilename, source, createdAt, updatedAt, educations, skills, projects]
      properties:
        id: {type: integer, format: int64}
        version: {type: integer, format: int64}
        name: {type: string}
        cvDocumentId: {type: integer, format: int64}
        cvOriginalFilename: {type: string}
        headline: {type: string, nullable: true}
        summary: {type: string, nullable: true}
        yearsExperience: {type: number, multipleOf: 0.1, minimum: 0, maximum: 99.9, nullable: true}
        targetPosition: {type: string, nullable: true}
        seniorityLevel: {type: string, nullable: true}
        source: {type: string, enum: [AUTO_PARSED, USER_EDITED]}
        confirmedAt: {type: string, format: date-time, nullable: true}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        educations: {type: array, items: {$ref: '#/components/schemas/ProfileEducation'}}
        skills: {type: array, items: {$ref: '#/components/schemas/ProfileSkill'}}
        projects: {type: array, items: {$ref: '#/components/schemas/ProfileProject'}}
    ProfileUpdateRequest:
      type: object
      required: [version, name, educations, skills, projects]
      properties:
        version: {type: integer, format: int64, minimum: 0}
        name: {type: string, minLength: 1, maxLength: 150}
        headline: {type: string, maxLength: 255, nullable: true}
        summary: {type: string, maxLength: 5000, nullable: true}
        yearsExperience: {type: number, multipleOf: 0.1, minimum: 0, maximum: 99.9, nullable: true}
        targetPosition: {type: string, maxLength: 150, nullable: true}
        seniorityLevel: {type: string, maxLength: 30, nullable: true}
        educations: {type: array, maxItems: 20, items: {$ref: '#/components/schemas/ProfileEducation'}}
        skills: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/ProfileSkill'}}
        projects: {type: array, maxItems: 50, items: {$ref: '#/components/schemas/ProfileProject'}}
...
security:
  - bearerAuth: []